# complete-api-docs.json
{
    "openapi":  "3.0.0",
    "paths":  {
...
                  "/api/ai/pronunciation":  {
                                                "post":  {
                                                             "operationId":  "AIController_scorePronunciation",
                                                             "summary":  "Score pronunciation audio for a word or phrase",
                                                             "description":  "Analyzes audio pronunciation and provides scoring and feedback (currently placeholder implementation)",
                                                             "parameters":  [

                                                                            ],
                                                             "requestBody":  {
                                                                                 "required":  true,
                                                                                 "description":  "Audio data and text for pronunciation analysis including audio file and target word/phrase",
                                                                                 "content":  {
                                                                                                 "application/json":  {
                                                                                                                          "schema":  {
                                                                                                                                         "$ref":  "#/components/schemas/PronunciationDto"
                                                                                                                                     }
                                                                                                                      }
                                                                                             }
                                                                             },
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "Pronunciation scored successfully"
                                                                                       },
                                                                               "401":  {
                                                                                           "description":  "Unauthorized"
                                                                                       }
                                                                           },
                                                             "tags":  [
                                                                          "AI"
                                                                      ],
                                                             "security":  [
                                                                              {
                                                                                  "bearer":  [

                                                                                             ]
                                                                              }
                                                                          ]
                                                         }
                                            },
                  "/api/ai/example":  {
                                          "post":  {
                                                       "operationId":  "AIController_generateExample",
                                                       "summary":  "Generate example sentence for context/word",
                                                       "description":  "Uses AI to generate contextual example sentences for vocabulary learning",
                                                       "parameters":  [

                                                                      ],
                                                       "requestBody":  {
                                                                           "required":  true,
                                                                           "description":  "Context and parameters for example generation including target word, language, and difficulty level",
                                                                           "content":  {
                                                                                           "application/json":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/GenerateExampleDto"
                                                                                                                               }
                                                                                                                }
                                                                                       }
                                                                       },
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "Example sentence generated successfully"
                                                                                 },
                                                                         "401":  {
                                                                                     "description":  "Unauthorized"
                                                                                 }
                                                                     },
                                                       "tags":  [
                                                                    "AI"
                                                                ],
                                                       "security":  [
                                                                        {
                                                                            "bearer":  [

                                                                                       ]
                                                                        }
                                                                    ]
                                                   }
                                      },
                  "/api/ai/explanation":  {
                                              "post":  {
                                                           "operationId":  "AIController_generateExplanation",
                                                           "summary":  "Generate explanation for a word within context",
                                                           "description":  "Provides AI-generated explanations for words based on their contextual usage",
                                                           "parameters":  [

                                                                          ],
                                                           "requestBody":  {
                                                                               "required":  true,
                                                                               "description":  "Word and context information for generating detailed explanations including usage and meaning",
                                                                               "content":  {
                                                                                               "application/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/GenerateExplanationDto"
                                                                                                                                   }
                                                                                                                    }
                                                                                           }
                                                                           },
                                                           "responses":  {
                                                                             "200":  {
                                                                                         "description":  "Word explanation generated successfully"
                                                                                     },
                                                                             "401":  {
                                                                                         "description":  "Unauthorized"
                                                                                     }
                                                                         },
                                                           "tags":  [
                                                                        "AI"
                                                                    ],
                                                           "security":  [
                                                                            {
                                                                                "bearer":  [

                                                                                           ]
                                                                            }
                                                                        ]
                                                       }
                                          },
                  "/api/ai/synonyms":  {
                                           "post":  {
                                                        "operationId":  "AIController_generateSynonyms",
                                                        "summary":  "Generate synonyms for a word",
                                                        "description":  "Uses AI to generate contextually appropriate synonyms for vocabulary expansion",
                                                        "parameters":  [

                                                                       ],
                                                        "requestBody":  {
                                                                            "required":  true,
                                                                            "description":  "Target word and language for synonym generation including context for appropriate alternatives",
                                                                            "content":  {
                                                                                            "application/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/GenerateSynonymsDto"
                                                                                                                                }
                                                                                                                 }
                                                                                        }
                                                                        },
                                                        "responses":  {
                                                                          "200":  {
                                                                                      "description":  "Synonyms generated successfully"
                                                                                  },
                                                                          "401":  {
                                                                                      "description":  "Unauthorized"
                                                                                  }
                                                                      },
                                                        "tags":  [
                                                                     "AI"
                                                                 ],
                                                        "security":  [
                                                                         {
                                                                             "bearer":  [

                                                                                        ]
                                                                         }
                                                                     ]
                                                    }
                                       },
                  "/api/ai/scheduling":  {
                                             "post":  {
                                                          "operationId":  "AIController_adjustScheduling",
                                                          "summary":  "Get AI-based scheduling adjustment recommendation",
                                                          "description":  "Provides AI-powered recommendations for optimizing spaced repetition scheduling based on user performance",
                                                          "parameters":  [

                                                                         ],
                                                          "requestBody":  {
                                                                              "required":  true,
                                                                              "description":  "User performance data including accuracy, streak days, and study patterns for scheduling optimization",
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/AdjustSchedulingDto"
                                                                                                                                  }
                                                                                                                   }
                                                                                          }
                                                                          },
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "Scheduling recommendation generated successfully"
                                                                                    },
                                                                            "401":  {
                                                                                        "description":  "Unauthorized"
                                                                                    }
                                                                        },
                                                          "tags":  [
                                                                       "AI"
                                                                   ],
                                                          "security":  [
                                                                           {
                                                                               "bearer":  [

                                                                                          ]
                                                                           }
                                                                       ]
                                                      }
                                         },
                  "/api/ai/recommendation":  {
                                                 "post":  {
                                                              "operationId":  "AIController_getRecommendation",
                                                              "summary":  "Get personalized study recommendations",
                                                              "description":  "Generates AI-powered personalized study recommendations based on user learning patterns and preferences",
                                                              "parameters":  [

                                                                             ],
                                                              "requestBody":  {
                                                                                  "required":  true,
                                                                                  "description":  "User preferences and context for personalized study recommendations including learning goals and current progress",
                                                                                  "content":  {
                                                                                                  "application/json":  {
                                                                                                                           "schema":  {
                                                                                                                                          "$ref":  "#/components/schemas/RecommendationDto"
                                                                                                                                      }
                                                                                                                       }
                                                                                              }
                                                                              },
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "Recommendations generated successfully"
                                                                                        },
                                                                                "401":  {
                                                                                            "description":  "Unauthorized"
                                                                                        }
                                                                            },
                                                              "tags":  [
                                                                           "AI"
                                                                       ],
                                                              "security":  [
                                                                               {
                                                                                   "bearer":  [

                                                                                              ]
                                                                               }
                                                                           ]
                                                          }
                                             },
...
              },
    "info":  {
                 "title":  "Matrus API",
                 "description":  "AI-powered Leitner System API",
                 "version":  "1.0",
                 "contact":  {

                             }
             },
...
    "servers":  [

                ],
    "components":  {
                       "securitySchemes":  {
                                               "bearer":  {
                                                              "scheme":  "bearer",
                                                              "bearerFormat":  "JWT",
                                                              "type":  "http"
                                                          }
                                           },
                       "schemas":  {
...
                                       "PronunciationDto":  {
                                                                "type":  "object",
                                                                "properties":  {
                                                                                   "word":  {
                                                                                                "type":  "string",
                                                                                                "description":  "The word or phrase to score pronunciation for",
                                                                                                "example":  "pronunciation"
                                                                                            },
                                                                                   "audioBase64":  {
                                                                                                       "type":  "string",
                                                                                                       "description":  "Base64 encoded audio data of the pronunciation attempt",
                                                                                                       "example":  "data:audio/wav;base64,UklGRnoGAABXQVZFZm10IBAAAAABAAEA..."
                                                                                                   },
                                                                                   "language":  {
                                                                                                    "type":  "string",
                                                                                                    "description":  "Language of the word being pronounced",
                                                                                                    "example":  "English"
                                                                                                }
                                                                               },
                                                                "required":  [
                                                                                 "word",
                                                                                 "audioBase64"
                                                                             ]
                                                            },
...
                                   }
                   }
}